backend/requirements.txt
uvicorn[standard]
fastapi
pydantic
pdfplumber
python-multipart
sse-starlette
google-generativeai
python-dotenv